#79 fix: moved filter strategies from platform to api
